UI SetUp for Checkout Page and Structs for Card/Ticket type
UI is set up as a template where backend code can be filled in to populate needed areas. Event listener will be needed to be set up for the button to send confirmation email with confirmation number and details.

The Checkout.cs page has two structs that can be used for 4 types of tickets and  types of card brands.

The ticket and the card have a selection tool (Picker) so users can select their card and ticket type.